Рефакторинг устранение ошибок оптимизация

+++ kmotion_hkd1.py
@@ -1,7 +1,5 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-from __future__ import absolute_import, division, unicode_literals, print_function, generators
-
 """
 Checks the size of the images directory deleteing the oldest directorys first
 when 90% of max_size is reached. Responds to a SIGHUP by re-reading its
@@ -11,8 +9,8 @@ configuration. Checks the current kmotion software version every 24 hours.
 import time
 from core import logger, utils
 from multiprocessing import Process
-import os
 from core.config import Settings
+from pathlib import Path
 
 log = logger.getLogger('kmotion', logger.ERROR)
 www_logs = logger.getLogger('www_logs', logger.DEBUG)
@@ -28,7 +26,6 @@ class Kmotion_Hkd1(Process):
         self.images_dbase_dir = ''  # the 'root' directory of the images dbase
         self.kmotion_dir = kmotion_dir
         self.max_size = 0  # max size permitted for the images dbase
-        self.motion_log = os.path.join(self.kmotion_dir, 'www', 'motion_out')
         self.read_config()
 
     def sleep(self, timeout):
@@ -43,20 +40,8 @@ class Kmotion_Hkd1(Process):
     def read_config(self):
         config_main = Settings.get_instance(self.kmotion_dir).get('kmotion_rc')
         log.setLevel(min(config_main['log_level'], log.getEffectiveLevel()))
-        self.images_dbase_dir = config_main['images_dbase_dir']
+        self.images_dbase_dir = Path(config_main['images_dbase_dir'])
         self.max_size = config_main['images_dbase_limit_gb'] * 2 ** 30
-
-    def truncate_motion_logs(self):
-        try:
-            with open(self.motion_log, 'r+', encoding="utf-8") as f_obj:
-                events = f_obj.read().splitlines()
-                if len(events) > 500:  # truncate logs
-                    events = events[-500:]
-                    f_obj.seek(0)
-                    f_obj.write('\n'.join(events))
-                    f_obj.truncate()
-        except Exception:
-            log.exception('truncate motion logs error')
 
     def run(self):
         """
@@ -67,43 +52,39 @@ class Kmotion_Hkd1(Process):
         return  : none
         """
         self.active = True
-        log.info('starting daemon [{pid}]'.format(pid=self.pid))
+        log.info(f'starting daemon [{self.pid}]')
         while self.active:
             try:
                 self.read_config()
 
                 while self.sleep(15 * 60):
-                    self.truncate_motion_logs()
-
                     # if > 90% of max_size_gb, delete oldest
                     _size = utils.get_dir_size(self.images_dbase_dir)
-                    log.info('size of {} = {}'.format(self.images_dbase_dir, utils.sizeof_fmt(_size)))
+                    log.info(f'size of {self.images_dbase_dir} = {utils.sizeof_fmt(_size)}')
                     if _size > self.max_size:
                         log.warn('image storage limit reached')
 
-                        dir_ = os.listdir(self.images_dbase_dir)
-                        dir_.sort()
-
                         # if need to delete current recording, shut down kmotion
-                        for d in dir_:
+                        for fulld in self.images_dbase_dir.glob('*'):
                             try:
-                                fulld = os.path.join(self.images_dbase_dir, d)
-                                if time.strftime('%Y%m%d') == d:
-                                    log.error('** CRITICAL ERROR ** crash - delete todays data, \'images_dbase\' is too small')
-                                    www_logs.error('Deleting todays data, \'images_dbase\' is too small')
+                                if fulld.is_dir():
+                                    d = fulld.name
+                                    if time.strftime('%Y%m%d') == d:
+                                        log.error("** CRITICAL ERROR ** crash - delete todays data, 'images_dbase' is too small")
+                                        www_logs.error("Deleting todays data, 'images_dbase' is too small")
 
-                                log.info('try to delete {dir}'.format(dir=fulld))
-                                if utils.rmdir(fulld):
-                                    log.warn(f'Deleting archive data for {d[:4]}/{d[4:6]}/{d[6:8]}')
-                                    www_logs.warn(f'Deleting archive data for {d[:4]}/{d[4:6]}/{d[6:8]}')
-                                    break
+                                    log.info(f'try to delete {fulld}')
+                                    if utils.rmdir(fulld):
+                                        log.warn(f'Deleting archive data for {d[:4]}/{d[4:6]}/{d[6:8]}')
+                                        www_logs.warn(f'Deleting archive data for {d[:4]}/{d[4:6]}/{d[6:8]}')
+                                        break
                             except Exception as e:
-                                log.error('deleting of "{dir}" error: {error}'.format(dir=fulld, error=e))
+                                log.error(f'deleting of "{fulld}" error: {e}')
 
             except Exception as e:  # global exception catch
                 log.critical('** CRITICAL ERROR **', exc_info=1)
                 self.sleep(60)
 
     def stop(self):
-        log.info('stop {name}'.format(name=__name__))
+        log.info(f'stop {__name__}')
         self.active = False
